# #451 - Revise type parameters of LinearMap