--- a/test/unit_LinearMap.jl
+++ b/test/unit_LinearMap.jl
@@ -58,12 +58,9 @@ for N in [Float64, Rational{Int}, Float32]
     @test an_element(lm) ∈ lm
 
     # check linear map between vector and set
-    X = Interval([0.9, 1.10445])
-    a = [-1., 2.]
-    @test a * X isa LinearMap
     X = BallInf(N[1.0], N(1.10445))
-    a = N[-1, 2.]
-    @test a * X isa LinearMap{N, N}
+    a = N[-1., 2.]
+    @test a * X isa LinearMap{N, BallInf{N}, N, Matrix{N}}
 
     # linear map with a ZeroSet
     X = N[0. -1. ; 1. 0.] * ZeroSet{N}(2)