Closes #451.

* [x] reorder type parameters
* [x] add type parameter for set
* [x] add type parameter for matrix