@schillic 2018-08-04
Done.
No sign of performance changes in `Reachability`. The  table shows the result of one run.

| model | `master` | this branch |
| -- | -- | -- |
| Heat (dense) | 18.47 s | 1.37 s |
| ISS (sparse) | 18.01 s | 1.54 s |