@schillic 2018-08-05
> hmm another possibility is to run https://github.com/JuliaReach/ARCH2018_RE

The package crashes for me right at the start with up-to-date versions (`Pkg.update()` gives no changes):

```julia
$ ./run.sh 
ERROR: MethodError: no method matching joinpath(::Void)
Closest candidates are:
  joinpath(::String, ::String) at path.jl:208
  joinpath(::AbstractString) at path.jl:192
  joinpath(::AbstractString, ::AbstractString) at path.jl:217
  ...
Stacktrace:
 [1] isfile(::Void) at ./stat.jl:279
 [2] #benchmarkpkg#26(::Void, ::Void, ::Bool, ::String, ::Function, ::String, ::PkgBenchmark.BenchmarkConfig) at .../.julia/v0.6/PkgBenchmark/src/runbenchmark.jl:50
 [3] benchmarkpkg(::String) at .../.julia/v0.6/PkgBenchmark/src/runbenchmark.jl:42
```
Can you confirm?