@schillic 2018-08-05
Good idea with the benchmark. There was actually an issue with the lazy input setting. We have to merge [`Reachability.jl`'s #227](https://github.com/JuliaReach/Reachability.jl/pull/227) after merging this PR.