@schillic 2018-08-05
Here are the results. The only major divergence is in the third to last row of the tables. Not sure what this means.

`master`:

| ID                                              | time            | GC time    | memory          | allocations |
|-------------------------------------------------|----------------:|-----------:|----------------:|------------:|
| `["Build", "(\"BLDC01-BDS01\", \"dense\")"]`    |    3.499 s (5%) | 613.745 ms |   2.02 GiB (1%) |    34402794 |
| `["Build", "(\"BLDC01-BDS01\", \"discrete\")"]` |    3.602 s (5%) | 648.678 ms |   2.02 GiB (1%) |    34401914 |
| `["Build", "(\"BLDF01-BDS01\", \"dense\")"]`    |    4.195 s (5%) | 665.526 ms |   2.13 GiB (1%) |    34604482 |
| `["Build", "(\"BLDF01-BDS01\", \"discrete\")"]` |    3.460 s (5%) | 586.382 ms |   2.04 GiB (1%) |    34445687 |
| `["ISS", "(\"ISS01\", \"dense\")"]`             |    5.329 s (5%) | 594.810 ms |   3.35 GiB (1%) |     4301237 |
| `["ISS", "(\"ISS01\", \"discrete\")"]`          |    1.150 s (5%) | 225.803 ms | 486.03 MiB (1%) |      494901 |
| `["ISS", "(\"ISS02\", \"dense\")"]`             | 411.301 ms (5%) |  44.137 ms | 370.63 MiB (1%) |      309096 |
| `["ISS", "(\"ISS02\", \"discrete\")"]`          | 378.490 ms (5%) |  38.753 ms | 363.26 MiB (1%) |      302548 |
| `["ISS", "(\"ISU01\", \"dense\")"]`             |    2.406 s (5%) | 237.859 ms |   1.25 GiB (1%) |     3925941 |
| `["ISS", "(\"ISU01\", \"discrete\")"]`          |    2.510 s (5%) | 221.900 ms | 937.63 MiB (1%) |     7081068 |
| `["ISS", "(\"ISU02\", \"dense\")"]`             | 299.624 ms (5%) |  37.902 ms | 140.66 MiB (1%) |     1778351 |
| `["ISS", "(\"ISU02\", \"discrete\")"]`          | 263.476 ms (5%) |  33.532 ms | 134.37 MiB (1%) |     1790951 |

this branch:

| ID                                              | time            | GC time    | memory          | allocations |
|-------------------------------------------------|----------------:|-----------:|----------------:|------------:|
| `["Build", "(\"BLDC01-BDS01\", \"dense\")"]`    |    2.933 s (5%) | 558.982 ms |   2.04 GiB (1%) |    34402452 |
| `["Build", "(\"BLDC01-BDS01\", \"discrete\")"]` |    3.090 s (5%) | 583.579 ms |   2.04 GiB (1%) |    34401816 |
| `["Build", "(\"BLDF01-BDS01\", \"dense\")"]`    |    3.091 s (5%) | 570.397 ms |   2.14 GiB (1%) |    34558934 |
| `["Build", "(\"BLDF01-BDS01\", \"discrete\")"]` |    3.081 s (5%) | 565.026 ms |   2.05 GiB (1%) |    34401151 |
| `["ISS", "(\"ISS01\", \"dense\")"]`             |    5.169 s (5%) | 600.321 ms |   3.35 GiB (1%) |     4437730 |
| `["ISS", "(\"ISS01\", \"discrete\")"]`          |    1.177 s (5%) |  67.154 ms | 543.93 MiB (1%) |      534890 |
| `["ISS", "(\"ISS02\", \"dense\")"]`             | 418.454 ms (5%) |  48.488 ms | 370.97 MiB (1%) |      315474 |
| `["ISS", "(\"ISS02\", \"discrete\")"]`          | 336.729 ms (5%) |  40.531 ms | 363.63 MiB (1%) |      310546 |
| `["ISS", "(\"ISU01\", \"dense\")"]`             |    2.213 s (5%) | 211.107 ms |   1.25 GiB (1%) |     3880448 |
| `["ISS", "(\"ISU01\", \"discrete\")"]`          |    9.659 s (5%) | 531.541 ms |   2.27 GiB (1%) |     7821270 |
| `["ISS", "(\"ISU02\", \"dense\")"]`             | 241.375 ms (5%) |  34.686 ms | 141.64 MiB (1%) |     1776713 |
| `["ISS", "(\"ISU02\", \"discrete\")"]`          | 207.377 ms (5%) |  32.919 ms | 135.38 MiB (1%) |     1790405 |